+++ packages/plonk/src/index.ts
@@ -1,1 +1,1 @@
-export * from "./maths";
+export * from "./math";
